Refactor LocationComparisonBanner
