+++ components/Tags/index.ts
@@ -1,1 +1,1 @@
-export * from '../Tags';
+export * from './Tags';
